+++ op_env/_cli.py
@@ -5,18 +5,12 @@ import os
 import pipes
 import subprocess
 import sys
-from typing import Any, Dict, List, Optional, Sequence, Union
+from typing import Any, List, Optional, Sequence, Union
 
 from typing_extensions import TypedDict
 import yaml
 
-from .op import (
-    _op_consolidated_fields,
-    _op_get_item,
-    _op_list_items,
-    _op_pluck_correct_field,
-    EnvVarName,
-)
+from .op import do_smart_lookups, EnvVarName
 
 
 class Arguments(TypedDict):
@@ -89,19 +83,6 @@ def parse_argv(argv: List[str]) -> Arguments:
                                       description=sh_desc)
     add_environment_arguments(sh_parser)
     return vars(parser.parse_args(argv[1:]))  # type: ignore
-
-
-# TODO: move to op.py?
-def do_smart_lookups(env_var_names: List[EnvVarName]) -> Dict[str, str]:
-    list_items_output = _op_list_items(env_var_names)
-    all_fields_to_seek = _op_consolidated_fields(env_var_names)
-    field_values_for_envvars = _op_get_item(list_items_output,
-                                            env_var_names,
-                                            all_fields_to_seek)
-    return {
-        env_var_name: _op_pluck_correct_field(env_var_name, field_values_for_envvars[env_var_name])
-        for env_var_name in field_values_for_envvars
-    }
 
 
 def process_args(args: Arguments) -> int:
